+++ AltCover.Fake/Fake.fs
@@ -1,8 +1,11 @@
+#if RUNNER
 namespace AltCover.Fake
 
 open System
+open System.IO
 open System.Reflection
 open AltCover
+open AltCover.DotNet
 
 module Trace =
   open Fake.Core
@@ -18,6 +21,11 @@ module Trace =
     | Some logging -> logging
     | None -> Default
 
+[<NoComparison>]
+type Implementation =
+  | DotNetCore
+  | Framework
+
 type Api =
   static member Prepare(args : PrepareParams, ?log : Logging) =
     AltCover.Api.Prepare args (Trace.DoDefault log)
@@ -25,13 +33,40 @@ type Api =
     AltCover.Api.Collect args (Trace.DoDefault log)
   static member Ipmo() = AltCover.Api.Ipmo()
   static member Version() = AltCover.Api.Version()
+  // Finds the tool from within the .nuget package
+  static member toolPath toolType =
+    let here = Assembly.GetExecutingAssembly().Location
+    let root = Path.Combine(Path.GetDirectoryName here, "../..")
+
+    let target =
+      match toolType with
+      | Framework _ -> "AltCover.exe"
+      | _ -> "AltCover.dll"
+    match Directory.GetFiles(root, target, SearchOption.AllDirectories)
+          |> Seq.filter (fun f ->
+               let coretype =
+                 f
+                 |> Path.GetDirectoryName
+                 |> Path.GetFileName
+               coretype.StartsWith("netstandard", StringComparison.Ordinal) |> not)
+          |> Seq.tryHead with
+    | Some path -> path
+    | None -> String.Empty
+#else
+namespace AltCover_Fake.DotNet
+
+open System
+open System.Reflection
+open AltCover.Internals
+open AltCover_Fake.DotNet.Testing
+#endif
 
 module DotNet =
   open Fake.DotNet
 
   type DotNet.TestOptions with
-    member self.WithParameters (prepare : PrepareParams) (collect : CollectParams) =
-      let options = AltCover.DotNet.ToTestArguments prepare collect
+
+    member private self.ExtendCustomParams options =
       let custom = self.Common.CustomParams
 
       let extended =
@@ -79,8 +114,8 @@ module DotNet =
       self.Common.GetType().GetFields(BindingFlags.NonPublic ||| BindingFlags.Instance)
       |> Array.iter (fun f ->
            f.SetValue(common,
-                      if f.Name <> "CustomParams@" then f.GetValue self.Common
-                      else extended :> obj))
+                      (if f.Name <> "CustomParams@" then f.GetValue self.Common
+                       else extended :> obj)))
       let TestOptionsConstructor = self.GetType().GetConstructors().[0]
 
       let args' =
@@ -94,6 +129,17 @@ module DotNet =
       self.GetType().GetFields(BindingFlags.NonPublic ||| BindingFlags.Instance)
       |> Array.iter (fun f ->
            f.SetValue(result,
-                      if f.Name <> "Common@" then f.GetValue self
-                      else common))
+                      (if f.Name <> "Common@" then f.GetValue self
+                       else common)))
       result :?> DotNet.TestOptions
+#if RUNNER
+
+    member self.WithParameters (prepare : PrepareParams) (collect : CollectParams) =
+#else
+    member self.WithParameters (prepare : AltCover.PrepareParams)
+           (collect : AltCover.CollectParams) =
+#endif
+
+      DotNet.ToTestArguments prepare collect |> self.ExtendCustomParams
+    member self.WithImportModule() = self.ExtendCustomParams "/p:AltCoverIpmo=true"
+    member self.WithGetVersion() = self.ExtendCustomParams "/p:AltCoverGetVersion=true"